+++ packages/caelus/src/index.ts
@@ -15,3 +15,4 @@ export * from "./astrocartography.js";
 export * from "./ephemeris.js";
 export * from "./features.js";
 export * from "./compiler.js";
+export * from "./lots.js";